refactor: expose compiler module
