+++ .gitignore
@@ -16,3 +16,4 @@ private
 build/
 monocore/build/
 diff
+pr.txt
