+++ monobase/lib/compiler/lexer/token.rs
@@ -1,4 +1,4 @@
-use crate::Span;
+use crate::compiler::Span;
 
 //--------------------------------------------------------------------------------------------------
 // Types
