+++ monobase/lib/compiler/parser/parser.rs
@@ -1,4 +1,4 @@
-use crate::Lexer;
+use crate::compiler::Lexer;
 
 //--------------------------------------------------------------------------------------------------
 // Exports
