+++ monobase/lib/lib.rs
@@ -3,10 +3,8 @@
 #![warn(missing_docs)]
 #![allow(clippy::module_inception)]
 
-mod compiler;
-
 //--------------------------------------------------------------------------------------------------
 // Exports
 //--------------------------------------------------------------------------------------------------
 
-pub use compiler::*;
+pub mod compiler;
